+++ controllers/agents.js
@@ -23,6 +23,7 @@ var router = require('express').Router();
  * @apiParam {String} [sort] Sorts the collection by a field or fields (separated by comma). Use +/- at the beginning to list in ascending or descending order.
  * @apiParam {String} [search] Looks for elements with the specified string.
  * @apiParam {String} [select] List of selected fields.
+ * @apiParam {String="active", "pending", "neverconnected", "disconnected"} [status] Filters by agent status. Use commas to enter multiple statuses.
  * @apiParam {String} [q] Query to filter results by. For example q="status=Active"
  * @apiParam {String} [older_than] Filters out disconnected agents for longer than specified. Time in seconds, '[n_days]d', '[n_hours]h', '[n_minutes]m' or '[n_seconds]s'. For never connected agents, uses the register date.
  * @apiParam {String} [os.platform] Filters by OS platform.
@@ -41,12 +42,12 @@ var router = require('express').Router();
  *
  */
 router.get('/', cache(), function(req, res) {
-    var extra_filters = {'status':'alphanumeric_param', 'os.platform':'alphanumeric_param',
+    var query_checks = {'status':'alphanumeric_param', 'os.platform':'alphanumeric_param',
                          'os.version':'alphanumeric_param', 'manager':'alphanumeric_param',
                          'version':'alphanumeric_param', 'node': 'alphanumeric_param',
                          'older_than':'timeframe_type', 'group':'alphanumeric_param',
                          'name': 'alphanumeric_param', 'ip': 'ips' };
-    templates.array_request("/agents", req, res, "agents", {}, {}, {}, extra_filters);
+    templates.array_request("/agents", req, res, "agents", {}, query_checks);
 })
 
 /**
@@ -110,8 +111,8 @@ router.get('/summary/os', cache(), function(req, res) {
  *
  */
 router.get('/no_group', cache(), function (req, res) {
-    extra_filters = {'status': 'alphanumeric_param'}
-    templates.array_request('/agents/no_group', req, res, "agents", {}, {}, {}, extra_filters);
+    query_checks = {'status': 'alphanumeric_param'}
+    templates.array_request('/agents/no_group', req, res, "agents", {}, query_checks);
 })
 
 /**
@@ -178,11 +179,10 @@ router.get('/groups', cache(), function(req, res) {
  *
  */
 router.get('/groups/:group_id', cache(), function(req, res) {
-    req_arguments = {'group_id': req.params.group_id};
     param_cheks = {'group_id':'names'};
-    extra_filters = {'status': 'alphanumeric_param'}
+    query_checks = {'status': 'alphanumeric_param'}
 
-    templates.array_request('/agents/groups/:group_id', req, res, "agents", req_arguments, param_cheks, {}, extra_filters);
+    templates.array_request('/agents/groups/:group_id', req, res, "agents", param_cheks, query_checks);
 });
 
 
@@ -1008,14 +1008,12 @@ router.post('/insert', function(req, res) {
  * @apiDescription Returns all the different combinations that agents have for the selected fields. It also indicates the total number of agents that have each combination.
  *
  * @apiExample {curl} Example usage:
- *     curl -u foo:bar -k -X GET "https://127.0.0.1:55000/agents/stats/distinct?pretty"
+ *     curl -u foo:bar -k -X GET "https://127.0.0.1:55000/agents/stats/distinct?pretty&fields=os.platform"
  *
  */
 router.get('/stats/distinct', cache(), function (req, res) {
-    req_arguments = {'fields': req.query.fields};
-    extra_query_params = {'fields':'select_param'};
-
-    templates.array_request('/agents/stats/distinct', req, res, "agents", req_arguments, {}, extra_query_params);
+    query_checks = {'fields':'select_param'};
+    templates.array_request('/agents/stats/distinct', req, res, "agents", {}, query_checks);
 })
 
 module.exports = router;