+++ controllers/syscheck.js
@@ -38,12 +38,9 @@ var router = require('express').Router();
  *
  */
 router.get('/:agent_id', cache(), function(req, res) {
-    extra_filters = {'event':'names', 'file':'paths', 'filetype':'names', 'md5':'hashes', 'sha1':'hashes', 'hash':'hashes'};
+    query_checks = {'event':'names', 'file':'paths', 'filetype':'names', 'md5':'hashes', 'sha1':'hashes', 'hash':'hashes', 'summary': 'names'};
     templates.array_request("/syscheck/:agent_id", req, res, "syscheck",
-                           {'agent_id': req.params.agent_id},
-                           {'agent_id':'numbers'},
-                           {'summary':'names'},
-                            extra_filters);
+                           {'agent_id':'numbers'}, query_checks);
 })
 
 /**